+++ extra_visitors/protobuf3/ariles2/visitors/protobuf3/writer.h
@@ -14,15 +14,13 @@ namespace ariles2
 {
     namespace ns_protobuf3
     {
-        class ARILES2_VISIBILITY_ATTRIBUTE Writer : public visitor::Base<visitor::Visitor, visitor::Parameters>
+        class ARILES2_VISIBILITY_ATTRIBUTE Writer : public visitor::Base<Writer, visitor::Parameters>
         {
         public:
             using Parameters = visitor::Parameters;
 
 
         public:
-            using visitor::Base<visitor::Visitor, Parameters>::getDefaultParameters;
-
             template <class t_Ariles>
             const Parameters &getParameters(const t_Ariles &ariles_class) const
             {