+++ include/ariles2/visitors/serialization.h
@@ -145,21 +145,7 @@ namespace ariles2
         };
 
 
-        template <class t_Derived, class t_Parameters>
-        class ARILES2_VISIBILITY_ATTRIBUTE Base : public visitor::Base<visitor::GenericVisitor, t_Parameters>
-        {
-        public:
-            using Parameters = t_Parameters;
-
-
-        public:
-            using visitor::Base<visitor::GenericVisitor, t_Parameters>::getDefaultParameters;
-
-            template <class t_Ariles>
-            const t_Parameters &getParameters(const t_Ariles &ariles_class) const
-            {
-                return (ariles_class.arilesGetParameters(*static_cast<const t_Derived *>(this)));
-            }
-        };
+        template<class t_Derived, class t_Parameters>
+        using Base = visitor::Base<t_Derived, t_Parameters>;
     }  // namespace serialization
 }  // namespace ariles2