+++ include/ariles2/visitors/write.h
@@ -44,11 +44,6 @@ namespace ariles2
         template <class t_Derived, class t_Parameters>
         class ARILES2_VISIBILITY_ATTRIBUTE VisitorBase : public serialization::Base<t_Derived, t_Parameters>
         {
-        protected:
-            VisitorBase(){};
-            ~VisitorBase(){};
-
-
         public:
             /**
              * @brief open configuration file