+++ tests/api_v2/fixtures/014_flags.h
@@ -29,19 +29,19 @@ namespace ariles_tests
             typename t_Visitor::Writer writer(getWriterInitializer("configurable.cfg"));
             BOOST_CHECK_EQUAL(
                     configurable.getExpectedWriteParameters().sloppy_maps_,
-                    configurable.arilesGetParameters(writer.getWriter()).sloppy_maps_);
+                    configurable.arilesGetParameters(writer.template get<ariles2::write::Visitor>()).sloppy_maps_);
             BOOST_CHECK_EQUAL(
                     configurable.getExpectedWriteParameters().sloppy_pairs_,
-                    configurable.arilesGetParameters(writer.getWriter()).sloppy_pairs_);
+                    configurable.arilesGetParameters(writer.template get<ariles2::write::Visitor>()).sloppy_pairs_);
             ariles2::apply(writer, configurable);
 
             typename t_Visitor::Reader reader(getReaderInitializer("configurable.cfg"));
             BOOST_CHECK_EQUAL(
                     configurable.getExpectedReadParameters().sloppy_maps_,
-                    configurable.arilesGetParameters(reader.getReader()).sloppy_maps_);
+                    configurable.arilesGetParameters(reader.template get<ariles2::read::Visitor>()).sloppy_maps_);
             BOOST_CHECK_EQUAL(
                     configurable.getExpectedReadParameters().sloppy_pairs_,
-                    configurable.arilesGetParameters(reader.getReader()).sloppy_pairs_);
+                    configurable.arilesGetParameters(reader.template get<ariles2::read::Visitor>()).sloppy_pairs_);
         }
     };
 }  // namespace ariles_tests