+++ tests/api_v2/types/copy/MultiDOFJointTrajectory.h
@@ -31,19 +31,12 @@ namespace trajectory_msgs
         typedef std::vector<
                 std::basic_string<
                         char,
-                        std::char_traits<char>,
-                        typename ContainerAllocator::template rebind<char>::other>,
-                typename ContainerAllocator::template rebind<std::basic_string<
-                        char,
-                        std::char_traits<char>,
-                        typename ContainerAllocator::template rebind<char>::other>>::other>
+                        std::char_traits<char>>>
                 _joint_names_type;
         _joint_names_type joint_names;
 
         typedef std::vector<
-                ::trajectory_msgs::MultiDOFJointTrajectoryPoint_<ContainerAllocator>,
-                typename ContainerAllocator::template rebind<
-                        ::trajectory_msgs::MultiDOFJointTrajectoryPoint_<ContainerAllocator>>::other>
+                ::trajectory_msgs::MultiDOFJointTrajectoryPoint_<ContainerAllocator>>
                 _points_type;
         _points_type points;
     };  // struct MultiDOFJointTrajectory_