+++ tests/api_v2/types/copy/MultiDOFJointTrajectoryPoint.h
@@ -27,20 +27,17 @@ namespace trajectory_msgs
 
 
         typedef std::vector<
-                ::geometry_msgs::Transform_<ContainerAllocator>,
-                typename ContainerAllocator::template rebind<::geometry_msgs::Transform_<ContainerAllocator>>::other>
+                ::geometry_msgs::Transform_<ContainerAllocator>>
                 _transforms_type;
         _transforms_type transforms;
 
         typedef std::vector<
-                ::geometry_msgs::Twist_<ContainerAllocator>,
-                typename ContainerAllocator::template rebind<::geometry_msgs::Twist_<ContainerAllocator>>::other>
+                ::geometry_msgs::Twist_<ContainerAllocator>>
                 _velocities_type;
         _velocities_type velocities;
 
         typedef std::vector<
-                ::geometry_msgs::Twist_<ContainerAllocator>,
-                typename ContainerAllocator::template rebind<::geometry_msgs::Twist_<ContainerAllocator>>::other>
+                ::geometry_msgs::Twist_<ContainerAllocator>>
                 _accelerations_type;
         _accelerations_type accelerations;